refactor: ScoreCalculator의 private 함수를 따로 object로 분리
